Add tests, base.py
