+++ pgmpy/causal_discovery/__init__.py
@@ -1,3 +1,4 @@
+from pgmpy.causal_discovery.base import BaseCausalEstimator
 from pgmpy.causal_discovery.PC import PC
 
-__all__ = ["PC"]
+__all__ = ["BaseCausalEstimator", "PC"]
